#23 에서 수정한 엔티티에 맞게 후속 코드 수정입니다

+++ src/main/java/art/snail/naillian/backend/domain/nail/controller/NailSetController.java
@@ -21,7 +21,7 @@ public class NailSetController {
 
     @GetMapping("/{id}")
     public Mono<CommonResponse<NailSetEmbedDTO<NailImageUrlDTO>>> getNailSet(
-            @PathVariable("id") Long id
+            @PathVariable("id") Integer id
     ) {
         return nailService.getNailsBySetId(id)
                 .switchIfEmpty(Mono.error(new ReportableError(HttpStatus.NOT_FOUND, "네일 세트를 찾을 수 없습니다.")))
