+++ src/main/java/art/snail/naillian/backend/domain/nail/repository/NailGroupRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
-public interface NailGroupRepository extends ReactiveCrudRepository<NailGroup, Long> {
+public interface NailGroupRepository extends ReactiveCrudRepository<NailGroup, Integer> {
     Flux<NailGroup> findAllBy(Pageable page);
 
 }
